titanic-through-ai-eyes/requirements.txt
openai>=1.3.9
gradio>=4.0
python-dotenv
numpy<2venv\Scripts\activate